Add html-report-utils.ts with HTML built helpers for report.
