+++ packages/steps-s3-copy/lambda/summarise-copy-lambda/report_template.html
@@ -331,7 +331,7 @@
               <div class="text-secondary small mt-2">Copied folder tree:</div>
             </div>
             <div class="card-body">
-              <div class="small">{{TREE_HTML}}</div>
+              <div class="small">{{DESTINATION_TREE_BLOCK}}</div>
             </div>
           </div>
         </div>
@@ -341,7 +341,7 @@
           <div class="card shadow-sm">
             <div class="card-body">
               <h2 class="h5 mb-3">{{FILES_TABLE_TITLE}}</h2>
-              <div class="small">{{FILES_TABLE}}</div>
+              <div class="small">{{FILES_TABLE_BLOCK}}</div>
             </div>
           </div>
         </div>
@@ -351,7 +351,7 @@
           <div class="card shadow-sm">
             <div class="card-body">
               <h2 class="h5 mb-3">Cost Estimation</h2>
-              <div class="small">{{COST_HTML}}</div>
+              <div class="small">{{COST_ESTIMATION_BLOCK}}</div>
             </div>
           </div>
         </div>
